refactor: 수의사 Test 파일 이동 (#56)
