+++ src/test/java/domain/vet/service/VetServiceTest.java
@@ -1,4 +1,4 @@
-package domain.appointment.service;
+package domain.vet.service;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
